Remade category fetching not to use PS category class

+++ classes/Repository/LanguageRepository.php
@@ -30,11 +30,11 @@ class LanguageRepository
     /**
      * @param string $isoCode
      *
-     * @return false|string|null
+     * @return int
      */
     public function getLanguageIdByIsoCode($isoCode)
     {
-        return Language::getIdByIso($isoCode);
+        return (int) Language::getIdByIso($isoCode);
     }
 
     /**
